Creates @lg-tools/storybook
