+++ .storybook/manager.ts
@@ -1,5 +1,6 @@
 import { addons } from '@storybook/addons';
-import theme from './theme';
+
+import theme from '@lg-tools/storybook/theme';
 
 addons.setConfig({
   name: 'LeafyGreen UI',
