+++ package.json
@@ -52,24 +52,10 @@
     "@leafygreen-ui/lib": "^10.1.0",
     "@leafygreen-ui/testing-lib": "*",
     "@lg-tools/build": "^0.0.1",
+    "@lg-tools/storybook": "^0.0.1",
+    "@lg-tools/storybook-decorators": "^0.0.1",
     "@slack/web-api": "^6.4.0",
-    "@storybook/addon-a11y": "^7.0.11",
-    "@storybook/addon-actions": "^7.0.11",
-    "@storybook/addon-docs": "^7.0.11",
-    "@storybook/addon-essentials": "^7.0.11",
-    "@storybook/addon-interactions": "^7.0.20",
-    "@storybook/addon-links": "^7.0.11",
-    "@storybook/addon-mdx-gfm": "^7.0.11",
-    "@storybook/addons": "^7.0.11",
-    "@storybook/api": "^7.0.11",
-    "@storybook/components": "^7.0.11",
-    "@storybook/core-events": "^7.0.11",
-    "@storybook/jest": "^0.1.0",
-    "@storybook/react": "^7.0.11",
-    "@storybook/react-webpack5": "^7.0.11",
-    "@storybook/testing-library": "^0.1.0",
-    "@storybook/theming": "^7.0.11",
-    "@svgr/webpack": "^8.0.0",
+    
     "@tanstack/react-table": "^8.7.3",
     "@testing-library/dom": "^7.2.1",
     "@testing-library/jest-dom": "^5.4.0",
@@ -131,7 +117,7 @@
     "regenerator-runtime": "^0.13.5",
     "source-map-loader": "^0.2.4",
     "storybook": "^7.0.11",
-    "storybook-addon-performance": "^0.17.1",
+
     "stream-browserify": "^3.0.0",
     "ts-node": "^8.8.2",
     "turbo": "^1.10.3",
