+++ tools/storybook-decorators/package.json
@@ -2,8 +2,10 @@
   "name": "@lg-tools/storybook-decorators",
   "version": "0.0.1",
   "description": "Storybook decorators used in LeafyGreen repos",
-  "main": "dist/index.js",
-  "license": "MIT",  
+  "main": "./dist/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/index.d.ts",
+  "license": "Apache-2.0",
   "scripts": {
     "build": "rollup --config rollup.config.mjs",
     "tsc": "tsc --build tsconfig.json"
@@ -15,6 +17,7 @@
     "@leafygreen-ui/leafygreen-provider": "^3.1.3",
     "@leafygreen-ui/palette": "^4.0.4",
     "@leafygreen-ui/tokens": "^2.1.1",
+    "@lg-tools/build": "^0.0.1",
     "@storybook/react": "^7.0.11",
     "lodash": "^4.17.21"
   }
