+++ tools/storybook-decorators/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "@lg-tools/config/package.tsconfig.json",
+  "extends": "@lg-tools/build/package.tsconfig.json",
   "compilerOptions": {
    "declarationDir": "dist",
    "outDir": "dist",
