+++ tools/storybook/babel.config.js
@@ -1,0 +1,2 @@
+import babel from "@lg-tools/build/babel.config"
+export default babel
