+++ tools/storybook/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@lg-tools/storybook",
+  "version": "0.0.1",
+  "description": "Storybook config for LeafyGreen",
+  "main": "main.ts",
+  "license": "MIT",
+  "dependencies": {
+    "@storybook/addon-a11y": "^7.0.11",
+    "@storybook/addon-actions": "^7.0.11",
+    "@storybook/addon-docs": "^7.0.11",
+    "@storybook/addon-essentials": "^7.0.11",
+    "@storybook/addon-interactions": "^7.0.20",
+    "@storybook/addon-links": "^7.0.11",
+    "@storybook/addon-mdx-gfm": "^7.0.11",
+    "@storybook/addons": "^7.0.11",
+    "@storybook/api": "^7.0.11",
+    "@storybook/components": "^7.0.11",
+    "@storybook/core-events": "^7.0.11",
+    "@storybook/jest": "^0.1.0",
+    "@storybook/react": "^7.0.11",
+    "@storybook/react-webpack5": "^7.0.11",
+    "@storybook/testing-library": "^0.1.0",
+    "@storybook/theming": "^7.0.11",
+    "@leafygreen-ui/palette": "^4.0.4",
+    "@leafygreen-ui/tokens": "^2.1.1",
+    "@lg-tools/build": "^0.0.1",
+    "@svgr/webpack": "^8.0.1",
+    "storybook-addon-performance": "^0.17.1"
+  }
+}
